edge_device/requirements.txt
opencv-python
numpy
paho-mqtt
filterpy
scikit-image
lap